// src/Interfaces/IBook.ts
export type Genre =
  | 'FICTION'
  | 'NON_FICTION'
  | 'SCIENCE'
  | 'HISTORY'
  | 'BIOGRAPHY'
  | 'FANTASY';
export  interface IBook {
        _id: string
        image: string;
        title : string;
        author: string;
        genre: Genre;
        isbn : string;
        description?: string;
        copies: number;
        available : boolean;
}